+++ server/service/src/document/patient/mod.rs
@@ -8,8 +8,8 @@ use crate::ListResult;
 pub use self::query::*;
 mod query;
 pub use self::update::*;
-pub mod patient_document_update;
 pub mod patient_schema;
+pub mod patient_updated;
 mod update;
 pub use self::search::*;
 mod search;
